Rewrite INSTALL script
The INSTALL script has been refactored now to have a smaller set of
initial configuration requirements which it uses to bootstrap the
installation of the rest of the tooling. This is done by introducting a
new concept called an "installer" which lives in `/installers`. These
are scripts that are executed from the git repo after cloning.

Using these new installer scripts, we now install nvim (on both osx and
linux) and then use vim-plug to install plugins without nearly as many
dependencies.

Right now the installer scripts are idempotent so running install
multiple times should be non-destructive.